add git helper
